# openapi.yaml
openapi: 3.0.3
info:
  title: Recipe Manager API
  description: |
    REST API for creating, reading, updating, and deleting Markdown-based recipes.

    The API uses JSON throughout, validates all write payloads, and returns
    [RFC 7807](https://www.rfc-editor.org/rfc/rfc7807) Problem Details on errors.
  version: 0.0.1
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  contact:
    name: Recipe Manager
    url: https://github.com/patrick473/recipe-manager-backend
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.recipe-manager.example.com
    description: Production server
...
security:
  - bearerAuth: []
paths:
...
  /recipes:
    get:
      summary: List recipes
      description: >
        Returns a page of recipes matching the given filter, sort, and
        pagination parameters. `q` and `tags` are combined with AND semantics
        when both are present.
      operationId: listRecipes
      tags: [Recipes]
      security: []
      parameters:
        - name: q
          in: query
          required: false
          description: >
            Case-insensitive substring match against `title` OR
            `description`. An empty/absent value means "no filter".
          schema:
            type: string
          example: banana
        - name: tags
          in: query
          required: false
          style: form
          explode: false
          description: >
            Comma-separated tag list. Matches recipes having *any* of the
            given tags (OR semantics among themselves), AND-ed with `q` if
            both are present. An empty/absent value means "no filter".
          schema:
            type: array
            items:
              type: string
          example: [breakfast, quick]
        - name: sort
          in: query
          required: false
          description: >
            Sort field and direction as `field,dir`. Allowed fields: `title`,
            `prepTimeMinutes`, `cookTimeMinutes`, `createdAt`, `updatedAt`.
            `prepTimeMinutes`/`cookTimeMinutes` always sort `null` values
            last, regardless of direction. Defaults to `title,asc`.
          schema:
            type: string
            default: "title,asc"
          example: "title,asc"
        - name: page
          in: query
          required: false
          description: Zero-based page index. Defaults to `0`.
          schema:
            type: integer
            format: int32
            minimum: 0
            default: 0
          example: 0
        - name: size
          in: query
          required: false
          description: >
            Page size. Defaults to `20`; clamped (not rejected) to a maximum
            of `100`.
          schema:
            type: integer
            format: int32
            minimum: 1
            default: 20
          example: 20
      responses:
        "200":
          description: Success — a page of recipes matching the given parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecipePageResponse"
              example:
                content:
                  - id: 1
                    title: Classic Banana Bread
                    description: Moist and simple banana bread, ready in 1 hour
                    content: "## Ingredients\n- 3 ripe bananas\n\n## Steps\n1. Mash bananas."
                    tags: [breakfast, quick]
                    prepTimeMinutes: 10
                    cookTimeMinutes: 60
                    servings: 8
                    createdAt: "2024-01-15T09:00:00Z"
                    updatedAt: "2024-01-15T09:00:00Z"
                page: 0
                size: 20
                totalElements: 1
                totalPages: 1
        "400":
          description: The `sort` parameter references an unknown field
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
              example:
                type: "https://example.com/errors/invalid-sort-field"
                title: Invalid Sort Field
                status: 400
                detail: "Unknown sort field: bogus"

    post:
      summary: Create a recipe
      description: >
        Persists a new recipe and returns the created resource, including the
        auto-generated `id` and timestamps.
      operationId: createRecipe
      tags: [Recipes]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecipeRequest"
            example:
              title: Classic Banana Bread
              description: Moist and simple banana bread, ready in 1 hour
              content: "## Ingredients\n- 3 ripe bananas\n\n## Steps\n1. Mash bananas."
              tags: [breakfast, quick]
              prepTimeMinutes: 10
              cookTimeMinutes: 60
              servings: 8
      responses:
        "201":
          description: Recipe created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecipeResponse"
        "400":
          description: Validation failed — one or more required fields are missing or invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
              example:
                type: "https://example.com/errors/validation-failed"
                title: Validation Failed
                status: 400
                detail: Request validation failed
                errors:
                  title: "Title must not be blank"
        "401":
          description: Missing, invalid, or expired bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: Authenticated but not permitted to perform this action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
  /recipes/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: Recipe surrogate key
        schema:
          type: integer
          format: int64
          example: 1

    get:
      summary: Get a single recipe
      description: Returns the recipe with the given `id`.
      operationId: getRecipe
      tags: [Recipes]
      security: []
      responses:
        "200":
          description: Recipe found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecipeResponse"
        "404":
          description: No recipe exists with the given `id`
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
              example:
                type: "https://example.com/errors/recipe-not-found"
                title: Recipe Not Found
                status: 404
                detail: "Recipe not found: 99"

    put:
      summary: Update a recipe
      description: >
        Replaces all mutable fields (`title`, `description`, `content`) on an
        existing recipe. The `id`, `createdAt`, and `updatedAt` fields are
        managed by the server and cannot be set by the caller.
      operationId: updateRecipe
      tags: [Recipes]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecipeRequest"
      responses:
        "200":
          description: Recipe updated — returns the full updated resource
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecipeResponse"
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: No recipe exists with the given `id`
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: Missing, invalid, or expired bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: Authenticated but not permitted to perform this action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"

    delete:
      summary: Delete a recipe
      description: >
        Permanently removes the recipe. Returns 204 on success. Returns 404 if
        no recipe with the given `id` exists.
      operationId: deleteRecipe
      tags: [Recipes]
      responses:
        "204":
          description: Recipe deleted — no content
        "404":
          description: No recipe exists with the given `id`
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: Missing, invalid, or expired bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: Authenticated but not permitted to perform this action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
  /recipes/{id}/image:
    parameters:
      - name: id
        in: path
        required: true
        description: Recipe surrogate key
        schema:
          type: integer
          format: int64
          example: 1

    post:
      summary: Upload a recipe's hero image
      description: >
        Uploads (or replaces) the recipe's hero image. A recipe has at most
        one image; uploading a new one deletes the previously stored file.
        Accepts `image/jpeg`, `image/png`, or `image/webp`, up to 5MB;
        content is sniffed server-side and rejected if it doesn't actually
        decode as an image of the declared type.
      operationId: uploadRecipeImage
      tags: [Recipes]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: The image file (jpeg, png, or webp)
      responses:
        "200":
          description: Image uploaded — returns the updated recipe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecipeResponse"
        "400":
          description: >
            Invalid upload — wrong content type, content that doesn't decode
            as an image, or file too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: No recipe exists with the given `id`
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: Missing, invalid, or expired bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: Authenticated but not permitted to perform this action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"

    delete:
      summary: Delete a recipe's hero image
      description: >
        Removes the recipe's stored image and clears its image reference.
        Idempotent — returns 200 even if the recipe had no image.
      operationId: deleteRecipeImage
      tags: [Recipes]
      responses:
        "200":
          description: Image removed (or already absent) — returns the updated recipe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecipeResponse"
        "404":
          description: No recipe exists with the given `id`
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: Missing, invalid, or expired bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: Authenticated but not permitted to perform this action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"

    get:
      summary: Get a recipe's hero image
      description: >
        Streams the raw image bytes for the recipe's hero image, with a
        long-lived, immutable `Cache-Control` header (filenames are
        content-addressed, so a cache-bust on replace is automatic).
      operationId: getRecipeImage
      tags: [Recipes]
      security: []
      responses:
        "200":
          description: The recipe's image bytes
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
            image/webp:
              schema:
                type: string
                format: binary
        "404":
          description: No recipe exists with the given `id`, or it has no image
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Send the token returned by `POST /auth/register` or
        `POST /auth/login` as `Authorization: Bearer <token>`.
  schemas:
    RecipeRequest:
      type: object
      description: Payload for creating or replacing a recipe
      required:
        - title
        - content
      properties:
        title:
          type: string
          maxLength: 255
          description: Short human-readable title shown in the recipe list
          example: Classic Banana Bread
        description:
          type: string
          maxLength: 512
          nullable: true
          description: Optional one-line summary displayed below the title
          example: Moist and simple banana bread, ready in 1 hour
        content:
          type: string
          description: Full recipe body in Markdown format
          example: "## Ingredients\n- 3 ripe bananas\n\n## Steps\n1. Mash bananas."
        tags:
          type: array
          items:
            type: string
          maxItems: 20
          nullable: true
          description: Optional freeform labels for categorizing and filtering recipes
          example: [breakfast, quick]
        prepTimeMinutes:
          type: integer
          format: int32
          nullable: true
          minimum: 0
          description: Estimated preparation time in minutes
          example: 10
        cookTimeMinutes:
          type: integer
          format: int32
          nullable: true
          minimum: 0
          description: Estimated cooking time in minutes
          example: 60
        servings:
          type: integer
          format: int32
          nullable: true
          minimum: 0
          description: Number of servings this recipe yields
          example: 8
    RecipeResponse:
      type: object
      description: A recipe as returned by the API
      required:
        - id
        - title
        - description
        - content
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
          format: int64
          description: Auto-generated surrogate key
          example: 1
          readOnly: true
        title:
          type: string
          maxLength: 255
          description: Short human-readable title
          example: Classic Banana Bread
        description:
          type: string
          maxLength: 512
          nullable: true
          description: Optional one-line summary
          example: Moist and simple banana bread, ready in 1 hour
        content:
          type: string
          description: Full recipe body in Markdown format
          example: "## Ingredients\n- 3 ripe bananas\n\n## Steps\n1. Mash bananas."
        tags:
          type: array
          items:
            type: string
          maxItems: 20
          nullable: true
          description: Optional freeform labels for categorizing and filtering recipes
          example: [breakfast, quick]
        prepTimeMinutes:
          type: integer
          format: int32
          nullable: true
          minimum: 0
          description: Estimated preparation time in minutes
          example: 10
        cookTimeMinutes:
          type: integer
          format: int32
          nullable: true
          minimum: 0
          description: Estimated cooking time in minutes
          example: 60
        servings:
          type: integer
          format: int32
          nullable: true
          minimum: 0
          description: Number of servings this recipe yields
          example: 8
        imageUrl:
          type: string
          nullable: true
          description: >
            Relative URL to the recipe's hero image (`/recipes/{id}/image`),
            or `null` if no image has been uploaded.
          example: "/recipes/1/image"
          readOnly: true
        createdAt:
          type: string
          format: date-time
          description: UTC timestamp of first creation
          example: "2024-01-15T09:00:00Z"
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          description: UTC timestamp of last modification
          example: "2024-01-15T10:30:00Z"
          readOnly: true
    RecipePageResponse:
      type: object
      description: A page of recipes matching the given filter/sort/pagination parameters
      required:
        - content
        - page
        - size
        - totalElements
        - totalPages
      properties:
        content:
          type: array
          description: The recipes on this page
          items:
            $ref: "#/components/schemas/RecipeResponse"
        page:
          type: integer
          format: int32
          description: Zero-based index of this page
          example: 0
        size:
          type: integer
          format: int32
          description: Maximum number of recipes per page
          example: 20
        totalElements:
          type: integer
          format: int64
          description: Total number of recipes matching the filter, across all pages
          example: 42
        totalPages:
          type: integer
          format: int32
          description: Total number of pages available
          example: 3
    ProblemDetail:
      type: object
      description: RFC 7807 Problem Details object returned on errors
      properties:
        type:
          type: string
          format: uri
          description: URI identifying the error type
          example: "https://example.com/errors/recipe-not-found"
        title:
          type: string
          description: Short, human-readable summary of the problem type
          example: Recipe Not Found
        status:
          type: integer
          description: HTTP status code
          example: 404
        detail:
          type: string
          description: Human-readable explanation of this specific occurrence
          example: "Recipe not found: 99"
        errors:
          type: object
          description: Field-level validation errors (only present on 400 responses)
          additionalProperties:
            type: string
          example:
            title: "Title must not be blank"
            content: "Content must not be blank"
...